machine-005: hot fixes
