+++ src/utils/apiMachine.js
@@ -1,5 +1,6 @@
 import { BASE_URL } from '../constants/url';
 import { AuthFormGlobalStore as formStore } from '../stores';
+import { getCookie } from './getCookie';
 
 function handleResponce(res) {
   if (res.ok) {
@@ -8,23 +9,12 @@ function handleResponce(res) {
   return Promise.reject(new Error(res.status));
 }
 
-export function getCookie(name) {
-  const matches = document.cookie.match(
-    new RegExp(
-      '(?:^|; )' +
-        name.replace(/([\.$?*|{}\(\)\[\]\\\/\+^])/g, '\\$1') +
-        '=([^;]*)'
-    )
-  );
-  return matches ? matches[1] : undefined;
-}
-
-export const getEncryption = (text, algorithm, key, is_encryption) => {
+export const postEncryption = (text, algorithm, key, is_encryption) => {
   const headers = () => {
     if (formStore.loggedIn) {
       return {
         'Content-Type': 'application/json',
-        Authorization: getCookie('auth._token.local'),
+        Authorization: `Bearer ${getCookie('access')}`,
       };
     }
     return { 'Content-Type': 'application/json' };
