+++ src/utils/getCookie.js
@@ -1,0 +1,10 @@
+export function getCookie(name) {
+  const matches = document.cookie.match(
+    new RegExp(
+      '(?:^|; )' +
+        name.replace(/([\.$?*|{}\(\)\[\]\\\/\+^])/g, '\\$1') +
+        '=([^;]*)'
+    )
+  );
+  return matches ? matches[1] : undefined;
+}
